requirements.txt
# Core dependencies
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
Pillow>=9.4.0
matplotlib>=3.6.0
opencv-python>=4.8.0
tqdm>=4.64.0

# SAM2 specific dependencies
hydra-core>=1.3.0
iopath>=0.1.9
pycocotools>=2.0.6

# Development and utilities
jupyter>=1.0.0
ipykernel>=6.20.0
ffmpeg-python>=0.2.0

# Optional GPU support (uncomment if CUDA available)
# torch>=2.1.0+cu118
# torchvision>=0.16.0+cu118